% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary.R
\name{calculate_summary}
\alias{calculate_summary}
\title{calculate_summary}
\usage{
calculate_summary(
  data,
  type,
  impute = FALSE,
  names.suffix = NULL,
  other.eval = FALSE
)
}
\arguments{
\item{data}{A data frame object, sample x feature}

\item{type}{A vector indicating sample types}

\item{impute}{If impute missing values}

\item{names.suffix}{Suffix to add on output variable names}

\item{other.eval}{If calculate other summary (MAD and VAR)}
}
\value{
A list of summaries for all types of samples
}
\description{
Calculate summary for subsets of dataset
}
