% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_t.R
\name{compute_t}
\alias{compute_t}
\title{compute_t}
\usage{
compute_t(
  data = NULL,
  formula = NULL,
  dv = NULL,
  output.file = NULL,
  map.names = NULL,
  scaling = FALSE
)
}
\arguments{
\item{data}{A data frame object, sample x feature.}

\item{formula}{A formula indicating independent variable. e.g."~ Group".}

\item{dv}{A vector of names of dependent variables for testing.}

\item{output.file}{Output file names.}

\item{map.names}{A vector for mapping between original and formatted variable
names.}
}
\value{
A data frame containing the t test result.
}
\description{
Compute t test.
}
\examples{
compute_t( data = data,
                     formula = "~ Group",
                     dv = dv_names,
                     map.names = NULL)

}
