% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks_summary.R
\name{filter_peaks_summary}
\alias{filter_peaks_summary}
\title{filter_peaks_summary}
\usage{
filter_peaks_summary(
  XCMSnExp = NULL,
  peaktable = NULL,
  sample.type = NULL,
  rsd.filter = NULL,
  rt.range = NULL,
  mean.filter = NULL
)
}
\arguments{
\item{XCMSnExp}{A XCMSnExp object. If null, peaktable should be mandatory}

\item{peaktable}{A dataframe of peaktable, feature x sample. If null, XCMSnExp
should be mandatory}

\item{sample.type}{A vector indicating sample types}

\item{rt.range}{Retention time range for keeping features}

\item{bl.thres}{Threshold for keeping features based on their values in one
type of sample compared to blanks. e.g. c( 3, "BL", "<", "PO")}

\item{rsd.po.thres}{Threshold for keeping features based on their rsd in pools.
e.g. c( "PO", "<", 40)}

\item{mean.po.thres}{Threshold for keeping features based on their mean in pools.
e.g. c( "PO", ">", 1000)}

\item{mean.thres}{Threshold for keeping features based on their mean values in
certain types. e.g. list( c( "PO100", ">", 1, "PO25"),
c( "PO25", ">", 1, "PO12.5") )}
}
\value{
A data frame object that contains filtered peaktable and summary statistics
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
