% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R, R/filter_peaks_summary.R
\name{mzrt_filter}
\alias{mzrt_filter}
\title{mzrt_filter}
\usage{
mzrt_filter(mzrt = NULL, index = NULL)

mzrt_filter(mzrt = NULL, index = NULL)
}
\arguments{
\item{mzrt}{A list of dataframe or vectors containing peaktable, feature.info.}

\item{index}{Index of features to remove.}
}
\value{
A list of filtered dataframe or vectors.

A list of filtered dataframe or vectors.
}
\description{
Apply the filter to all the elements in a list.

Apply the filter to all the elements in a list.
}
\examples{
mzrt_filter( mzrt = list,
                       index = index)

mzrt_filter( mzrt = list,
                       index = index)

}
