% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(
  datatable = NULL,
  filename = NULL,
  title = NULL,
  p.cut.off = 0.05,
  max.overlaps = 10,
  color.manual = c("#B4464B", "#4682B4", "grey50")
)
}
\arguments{
\item{filename}{Output file name.}

\item{title}{Title of the plot.}

\item{p.cut.off}{p value cut off.}

\item{max.overlaps}{Maximum number of overlapping labels.}

\item{color.manual}{Colors used for different groups.}

\item{A}{data frame used for plotting.}
}
\value{
A volcano plot.
}
\description{
Create volcano plot.
}
