% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_bubble.R
\name{plot_volcano_bubble}
\alias{plot_volcano_bubble}
\title{plot_volcano_bubble}
\usage{
plot_volcano_bubble(
  datatable = NULL,
  filename = NULL,
  title = NULL,
  p.cut.off = 0.05,
  max.overlaps = 10,
  color.manual = c("#B4464B", "#4682B4", "grey50", "#B4AF46"),
  ratio_range = c(-10, 2),
  bubble_size_range = c(-1, 6),
  breaks = c(-10, -7.5, -5, -2.5, 0, 2.5),
  bubble_name = "Cell/media ratio"
)
}
\arguments{
\item{filename}{Output file name.}

\item{title}{Title of the plot.}

\item{p.cut.off}{p value cut off.}

\item{max.overlaps}{Maximum number of overlapping labels.}

\item{color.manual}{Colors used for different groups.}

\item{ratio_range}{Range of cell/media ratio, e.g. c(-10, 2)}

\item{bubble_size_range}{Range of bubble size, e.g. c(-1, 6)}

\item{breaks}{Breaks for cell/media ratio}

\item{bubble_name}{Legend name for bubbles}

\item{A}{data frame used for plotting.}
}
\value{
A volcano plot with different bubble size.
}
\description{
Create volcano plot with different bubble size.
}
