% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_plot_permeation.R
\name{run_analysis_plot_permeation}
\alias{run_analysis_plot_permeation}
\title{run_analysis_plot}
\usage{
run_analysis_plot_permeation(
  data = NULL,
  media.thres = 500,
  FC_thres = 2,
  permeation.features = NULL,
  path.result = getwd(),
  p.cut.off = 0.05,
  max.overlaps = 10,
  map.names = NULL,
  feature.info = NULL,
  color.manual = c("#B4464B", "#4682B4", "grey50", "#B4AF46"),
  fig.format = ".svg",
  bubble = TRUE,
  ratio_range = c(-10, 2),
  bubble_size_range = c(-1, 6),
  breaks = c(-10, -7.5, -5, -2.5, 0, 2.5),
  bubble_name = "Cell/media ratio"
)
}
\arguments{
\item{media.thres}{A threshold to label if feature exist in the media.}

\item{FC_thres}{A fold change threshold for highlighting only feature with
FC >= FC_thres.}

\item{permeation.features}{A list of features to detect permeation.}

\item{path.result}{Path of the result folder.}

\item{p.cut.off}{p value cut off.}

\item{max.overlaps}{Maximum number of overlapping labels.}

\item{map.names}{A vector for mapping between original and formatted variable
names.}

\item{feature.info}{Feature information.}

\item{color.manual}{Colors used for different groups.}

\item{fig.format}{Output figure format}

\item{bubble}{Indicate if plot bubble or not, e.g. TRUE/FALSE}

\item{ratio_range}{Range of cell/media ratio, e.g. c(-10, 2)}

\item{bubble_size_range}{Range of bubble size, e.g. c(-1, 6)}

\item{breaks}{Breaks for cell/media ratio}

\item{bubble_name}{Legend name for bubbles}

\item{A}{data frame for analysis, sample x feature.}
}
\value{
t test results and volcano plot with different bubble size for each gene.
}
\description{
A wrapper function to perform t test to data and then generate
volcano plot with different bubble size.
}
