% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary.R
\name{calculate_summary}
\alias{calculate_summary}
\title{calculate_summary}
\usage{
calculate_summary(
  data,
  type,
  impute = FALSE,
  names.suffix = NULL,
  other.eval = FALSE
)
}
\arguments{
\item{data}{A data frame object, sample x feature.}

\item{type}{A vector indicating sample types.}

\item{impute}{TRUE/FALSE, if impute missing values, if TRUE, half minimum is imputed.}

\item{names.suffix}{Suffix to add to the output variable names.}

\item{other.eval}{TRUE/FALSE, if calculate other summary (MAD and VAR).}
}
\value{
A list of summaries for all types of samples.
}
\description{
Calculate summary for subsets of dataset.
}
\examples{
calculate_summary( data = data.norm,
                             type = sample.info.neg$Sample.type,
                             names.suffix = "norm")

}
