% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_t.R
\name{compute_t}
\alias{compute_t}
\title{compute_t}
\usage{
compute_t(
  data = NULL,
  formula = NULL,
  dv = NULL,
  output.file = NULL,
  map.names = NULL,
  scaling = FALSE
)
}
\arguments{
\item{data}{= NULL,}

\item{formula}{= NULL,}

\item{dv}{= NULL,}

\item{output.file}{= NULL,}

\item{map.names}{= NULL}
}
\value{
A dataframe
}
\description{
Compute t test
}
