% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks_summary.R
\name{filter_peaks_summary}
\alias{filter_peaks_summary}
\title{filter_peaks_summary}
\usage{
filter_peaks_summary(
  XCMSnExp = NULL,
  peaktable = NULL,
  sample.type = NULL,
  rsd.filter = NULL,
  rt.range = NULL,
  mean.filter = NULL
)
}
\arguments{
\item{XCMSnExp}{A XCMSnExp object. If null, peaktable should be mandatory.}

\item{peaktable}{A dataframe of peaktable, feature x sample. If null, XCMSnExp
should be mandatory.}

\item{sample.type}{A vector indicating sample types.}

\item{rsd.filter}{A list of threshold for keeping features based on their rsd in QCs.
e.g. c( 1, "QC", "<", 0.3).}

\item{rt.range}{Retention time range for keeping features. e.g. c(0.5, 30).}

\item{mean.filter}{A list of threshold for keeping features based on their mean in QCs
or other sample types. e.g. list( c( 1, "PO", ">", "1000"),
c( 1, "PO100", ">", "PO50") ).}
}
\value{
A data frame object that contains filtered peaktable and summary statistics.
}
\description{
Filter the peaks based on different criteria and generate summary
statistics.
}
\examples{
filter_peaks_summary( peaktable = peaks,
                                sample.type = sample.info.pos$Sample.type,
                                rsd.filter = c( 1, "QC", "<", 0.3),
                                mean.filter = c(1, "QC", ">", 500),
                                rt.range = c(0.5, 30))

}
