% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R
\name{mzrt_filter}
\alias{mzrt_filter}
\title{mzrt_filter}
\usage{
mzrt_filter(mzrt = NULL, index = NULL)
}
\arguments{
\item{mzrt}{A list of dataframe or vectors containing peaktable, feature.info}

\item{index}{Index of features to remove}
}
\value{
A list of filtered dataframe or vectors
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
