% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(
  datatable = NULL,
  filename = NULL,
  title = NULL,
  FC_thres = NULL,
  p.cut.off = NULL,
  max.overlaps = NULL,
  color.manual = c("#B4464B", "#4682B4", "grey50")
)
}
\arguments{
\item{title}{description}

\item{color.manual}{}
}
\value{
A data frame
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
