% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_bubble.R
\name{plot_volcano_bubble}
\alias{plot_volcano_bubble}
\title{plot_volcano_bubble}
\usage{
plot_volcano_bubble(
  datatable = NULL,
  filename = NULL,
  title = NULL,
  p.cut.off = 0.05,
  max.overlaps = 10,
  color.manual = c("#B4464B", "#4682B4", "grey50"),
  ratio_range = c(-10, 2),
  max_size = 6
)
}
\arguments{
\item{filename}{Output file name.}

\item{title}{Title of the plot.}

\item{p.cut.off}{p value cut off.}

\item{max.overlaps}{Maximum number of overlapping labels.}

\item{color.manual}{Colors used for different groups.}

\item{ratio_range}{Range of cell/media ratio}

\item{max_size}{Maximum bubble size}

\item{A}{data frame used for plotting.}
}
\value{
A volcano plot with different bubble size.
}
\description{
Create volcano plot with different bubble size.
}
