% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_plot.R
\name{run_analysis_plot}
\alias{run_analysis_plot}
\title{run_analysis_plot}
\usage{
run_analysis_plot(
  data = NULL,
  media.thres = 500,
  FC_thres = 2,
  path.result = getwd(),
  p.cut.off = 0.05,
  max.overlaps = 10,
  map.names = NULL,
  feature.info = NULL,
  color.manual = c("#B4464B", "#4682B4", "grey50")
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
