% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_plot.R
\name{run_analysis_plot}
\alias{run_analysis_plot}
\title{run_analysis_plot}
\usage{
run_analysis_plot(
  data = NULL,
  media.thres = 500,
  FC_thres = 2,
  path.result = getwd(),
  p.cut.off = 0.05,
  max.overlaps = 10,
  map.names = NULL,
  feature.info = NULL,
  color.manual = c("#B4464B", "#4682B4", "grey50")
)
}
\arguments{
\item{media.thres}{A threshold to label if feature exist in the media.}

\item{FC_thres}{A fold change threshold for highlighting only feature with
FC >= FC_thres.}

\item{path.result}{Path of the result folder.}

\item{p.cut.off}{p value cut off.}

\item{max.overlaps}{Maximum number of overlapping labels.}

\item{map.names}{A vector for mapping between original and formatted variable
names.}

\item{feature.info}{Feature information.}

\item{color.manual}{Colors used for different groups.}

\item{A}{data frame for analysis, sample x feature.}
}
\value{
t test results and volcano plot with different bubble size for each gene.
}
\description{
A wrapper function to perform t test to data and then generate
volcano plot with different bubble size.
}
